add_test

+++ database/migrations/2022_10_21_132538_create_roll_histories_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained('users');
             $table->integer('number');
-            $table->integer('reward');
+            $table->float('reward');
             $table->timestamps();
         });
     }
